requirements.txt
aiodns==2.0.0
aiohttp==3.5.4
amqp==2.4.2
arrow==0.14.0
async-timeout==3.0.1
attrs==19.1.0
billiard==3.6.0.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
croniter==0.3.30
Deprecated==1.2.5
Flask==1.0.3
Flask-CLI==0.4.0
Flask-RQ2==18.3
Flask-Script==2.0.6
gevent==1.4.0
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.5.0
MarkupSafe==1.1.1
multidict==4.5.2
pycares==3.0.0
pycparser==2.19
PyGithub==1.43.7
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.2
pytz==2019.1
redis==3.2.1
requests==2.22.0
rq==1.0
rq-dashboard==0.5.1
rq-scheduler==0.9
six==1.12.0
tabulate==0.8.3
urllib3==1.25.2
vine==1.3.0
websocket==0.2.1
websocket-client==0.56.0
Werkzeug==0.15.4
wrapt==1.11.1
yarl==1.3.0
